refactor: address PR review feedback on verify/verifyAll/verifyEach
- Extract IrErrorCollector (mirrors IrValueRecorder) so consumers pass an
  instance around and call irValidateCollectedErrors() instead of
  building the call inline.
- Move fresh-collector construction into ConditionRecorderDeclarations.kt
  as irCollectingErrorCollectorDeclaration, alongside the renamed
  irStaticErrorCollectorDeclaration (was irAmbientErrorCollectorDeclaration).
- Extract ConditionStatementsRewriter into its own file, since it's used
  by both ConditionRewriter and HelperMethodRewriter.
- Audit and trim comments that just restated the code.
- Add KDoc to verify/verifyAll/verifyEach modeled on Spock's own
  Specification.java javadoc.
- Fix smoke tests to always start with given/when instead of an
  unlabeled leading statement.
- Docs: drop the return@ and "use namer" asides from the limitations
  list, remove em dashes.

Document two new conventions in AGENTS.md: no em dashes in prose, and a
robot-emoji prefix when replying to PR comments on the user's behalf.
